Redefine options for gatsby-theme-events #374
