+++ .changeset/weak-lizards-kneel.md
@@ -1,0 +1,5 @@
+---
+'@undataforum/gatsby-theme-events': minor
+---
+
+Redefine options for gatsby-theme-events
