+++ site/content/events/multiple-paras-with-import.mdx
@@ -9,16 +9,16 @@ moderators:
 speakers:
   - nance-mcfadin
   - suzy-cahan
-registration: /registration
+registrationLink: /registration
 ---
 
-import { Box } from 'theme-ui'; import { Video } from '@undataforum/components';
+import { Box } from 'theme-ui';
+import { Video } from '@undataforum/components';
 
 You should see an entire paragraph in the preview. This event covers the test
-case **multiple paragraphs** with import statement **before** the first
-paragraph. This is a filler sentence to make the **first paragraph longer**. If
-you see this sentence in the preview and it is the last sentence, everything
-works fine.
+case **multiple paragraphs with import statement before the first paragraph**.
+This is a filler sentence to make the first paragraph longer. If you see this
+sentence in the preview and it is the last sentence, everything works fine.
 
 Anim do laborum fugiat laboris aliqua eu et sint minim amet consequat sit.
 Commodo irure eiusmod enim eu enim cupidatat proident ipsum mollit sint eu.
@@ -37,5 +37,9 @@ exercitation in ex elit.
 ## Recording
 
 <Box mb={3}>
-  <Video id="Vb0c_sIzoUw" title="Webinar: Quis laborum ipsum non mollit ex magna consectetur id amet voluptate" ratio={16/9} />
+  <Video
+    id="Vb0c_sIzoUw"
+    title="Webinar: Quis laborum ipsum non mollit ex magna consectetur id amet voluptate"
+    ratio={16 / 9}
+  />
 </Box>
