+++ site/content/events/one-para-with-import.mdx
@@ -15,6 +15,6 @@ speakers:
 import { Box } from 'theme-ui';
 
 You should see an entire paragraph in the preview. This event covers the test
-case one paragraph with import statement before the paragraph. This is a filler
-sentence to make the paragraph longer. If you see this sentence in the preview
-and it is the last sentence, everything works fine.
+case **one paragraph with import statement before the paragraph**. This is a
+filler sentence to make the paragraph longer. If you see this sentence in the
+preview and it is the last sentence, everything works fine.
