+++ site/content/events/one-para-without-import.md
@@ -2,14 +2,11 @@
 type: Webinar
 title: Ad labore ipsum elit pariatur tempor pariatur fugiat officia ullamco
 start: 2019-08-24T09:00
-end: 2019-08-25T12:00
+end: 2019-08-24T10:00
 zone: America/Los_Angeles
-speakers:
-  - sandra-abramovici-floris
-  - clifford-skipperbottom
 ---
 
 You should see an entire paragraph in the preview. This event covers the test
-case one paragraph without import statement before the paragraph. This is a
+case **one paragraph without import statement before the paragraph**. This is a
 filler sentence to make the paragraph longer. If you see this sentence in the
 preview and it is the last sentence, everything works fine.
