+++ site/gatsby-config.js
@@ -13,8 +13,6 @@ module.exports = {
       resolve: '@undataforum/gatsby-theme-events',
       options: {
         basePath: '/events',
-        description:
-          'Events in preparation of the 2020 United Nations World Data Forum.',
       },
     },
     {
