--- themes/gatsby-theme-events/defaultOptions.js
@@ -1,11 +1,0 @@
-// Default theme options.
-module.exports = {
-  assetPath: 'content/assets/events',
-  basePath: '',
-  contentPath: 'content/events',
-  description:
-    'Events page description placeholder (from @undataforum/gatsby-theme-events).',
-  profileType: 'profile',
-  title: 'Events',
-  type: 'event',
-};
