+++ themes/gatsby-theme-events/gatsby-config.js
@@ -1,35 +1,23 @@
-const defaultOptions = require('./defaultOptions');
+const withDefaults = require('./utils/default-options');
 
-module.exports = ({
-  assetPath = defaultOptions.assetPath,
-  basePath = defaultOptions.basePath,
-  contentPath = defaultOptions.contentPath,
-  type = defaultOptions.type,
-  profileType = defaultOptions.profileType,
-}) => ({
-  siteMetadata: {
-    basePath,
-  },
-  plugins: [
-    '@undataforum/gatsby-theme-base',
-    {
-      resolve: '@undataforum/gatsby-theme-profiles',
-      options: { basePath: '/profiles', type: profileType },
-    },
-    // You can Filter File nodes by sourceInstanceName:
-    // https://www.gatsbyjs.org/packages/gatsby-source-filesystem/#how-to-query
-    {
-      resolve: 'gatsby-source-filesystem',
-      options: {
-        path: contentPath,
-        name: type,
+module.exports = themeOptions => {
+  const { contentPath, assetPath, collection } = withDefaults(themeOptions);
+  return {
+    plugins: [
+      '@undataforum/gatsby-theme-base',
+      {
+        resolve: 'gatsby-source-filesystem',
+        options: {
+          path: contentPath,
+          name: collection,
+        },
       },
-    },
-    {
-      resolve: 'gatsby-source-filesystem',
-      options: {
-        path: assetPath,
+      {
+        resolve: 'gatsby-source-filesystem',
+        options: {
+          path: assetPath,
+        },
       },
-    },
-  ],
-});
+    ],
+  };
+};
