+++ themes/gatsby-theme-events/src/i18n/messages.js
@@ -1,0 +1,5 @@
+import en from './translations/en';
+
+const messages = { en };
+
+export default messages;
