+++ themes/gatsby-theme-events/src/i18n/translations/en.js
@@ -1,0 +1,10 @@
+// These are the keys for the default collection `events`.
+// Shadow this file to override these keys and to create new keys for your collections.
+const en = {
+  'events.title': 'Events',
+  'events.tag': 'Event',
+  'events.moderator': 'Moderator',
+  'events.description': 'Description for overview page of events collection.',
+};
+
+export default en;
