+++ themes/gatsby-theme-events/utils/default-options.js
@@ -1,0 +1,8 @@
+module.exports = options => {
+  const basePath = options.basePath || '/';
+  const contentPath = options.contentPath || 'content/events';
+  const assetPath = options.assetPath || 'content/assets';
+  const collection = options.collection || 'events';
+  const profiles = options.profiles || undefined;
+  return { basePath, contentPath, assetPath, collection, profiles };
+};
