+++ themes/gatsby-theme-events/utils/url-resolve.js
@@ -1,0 +1,7 @@
+const { urlResolve } = require(`gatsby-core-utils`);
+
+module.exports = (...segments) => {
+  const path = urlResolve(...segments);
+  // Add trailing slash if there is none.
+  return path.replace(/\/*$/, `/`);
+};
